tools/autolabel_ver1/requirement.txt
#conda create -n label python=3.7 -y
#conda activate label
jupyterlab==3.6.8
pillow==9.5.0


























